Kernel: Add system call interface and yield() syscall
Use SVC exception to build the system call interface. The system call
number is sent on the stack and SVC 0 is used for calling the exception.
yield() system call has been implemented with syscall number 1, it takes
the cpu from the running task and gives it to the next task on the list.

Signed-off-by: Puranjay Mohan <puranjay12@gmail.com>